move service code to multiuser_windows_service.go
