+++ multiuser_windows.go
@@ -14,93 +14,12 @@ import (
 	"strconv"
 	"strings"
 	"syscall"
-	"time"
 
 	"github.com/taskcluster/generic-worker/process"
 	"github.com/taskcluster/generic-worker/runtime"
 	gwruntime "github.com/taskcluster/generic-worker/runtime"
 	"github.com/taskcluster/generic-worker/win32"
-	"golang.org/x/sys/windows"
-	"golang.org/x/sys/windows/registry"
-	"golang.org/x/sys/windows/svc"
-	"golang.org/x/sys/windows/svc/debug"
-	"golang.org/x/sys/windows/svc/eventlog"
-	"golang.org/x/sys/windows/svc/mgr"
 )
-
-var elog debug.Log
-
-type windowsService struct{}
-
-// implements Execute for svc.Handler
-func (*windowsService) Execute(args []string, r <-chan svc.ChangeRequest, changes chan<- svc.Status) (bool, uint32) {
-	const cmdsAccepted = svc.AcceptStop | svc.AcceptShutdown | svc.AcceptPauseAndContinue
-	changes <- svc.Status{State: svc.StartPending}
-
-	// Start worker with interruptChan
-	interruptChan := make(chan os.Signal, 1)
-
-	go RunWorker(interruptChan)
-
-	changes <- svc.Status{State: svc.Running, Accepts: cmdsAccepted}
-loop:
-	for {
-		select {
-		case c := <-r:
-			switch c.Cmd {
-			case svc.Interrogate:
-				changes <- c.CurrentStatus
-				// Testing deadlock from https://code.google.com/p/winsvc/issues/detail?id=4
-				time.Sleep(100 * time.Millisecond)
-				changes <- c.CurrentStatus
-			case svc.Stop, svc.Shutdown:
-				log.Printf("Shutting down, received %v", c)
-				interruptChan <- os.Interrupt
-				break loop
-			case svc.Pause:
-				changes <- svc.Status{State: svc.Paused, Accepts: cmdsAccepted}
-			case svc.Continue:
-				changes <- svc.Status{State: svc.Running, Accepts: cmdsAccepted}
-			default:
-				log.Printf("Unexpected control request #%d", c)
-			}
-		}
-	}
-	changes <- svc.Status{State: svc.StopPending}
-	return false, 0
-}
-
-func runService(name string, isDebug bool) ExitCode {
-	var err error
-	if name == "" {
-		name = "Generic Worker"
-	}
-
-	if isDebug {
-		log.Printf("Debug mode enabled, not using eventlog.")
-		elog = debug.New(name)
-	} else {
-		elog, err = eventlog.Open(name)
-		if err != nil {
-			fmt.Printf("Could not open eventlog: %v", err)
-			return INTERNAL_ERROR
-		}
-	}
-	defer elog.Close()
-
-	elog.Info(1, fmt.Sprintf("Starting service %q", name))
-	run := svc.Run
-	if isDebug {
-		run = debug.Run
-	}
-	err = run(name, &windowsService{})
-	if err != nil {
-		elog.Error(1, fmt.Sprintf("Service %q failed: %v", name, err))
-		return INTERNAL_ERROR
-	}
-	elog.Info(1, fmt.Sprintf("Stopped service %q", name))
-	return 0
-}
 
 func (task *TaskRun) formatCommand(index int) string {
 	return task.Payload.Command[index]
@@ -322,89 +241,6 @@ func install(arguments map[string]interface{}) (err error) {
 	return fmt.Errorf("Unknown install target - only 'service' is allowed")
 }
 
-// deploys the generic worker as a windows service named name
-// running as the user LocalSystem
-// if the service already exists we skip.
-func deployService(configFile, name, exePath string, configureForAWS bool, configureForGCP bool) error {
-	m, err := mgr.Connect()
-	if err != nil {
-		return err
-	}
-	defer m.Disconnect()
-	s, err := m.OpenService(name)
-	if err == nil {
-		s.Close()
-		return fmt.Errorf("service %s already exists", name)
-	}
-	args := []string{
-		"run-service",
-	}
-	if configFile != "" {
-		args = append(args, "--config", configFile)
-	}
-	if configureForAWS {
-		args = append(args, "--configure-for-aws")
-	}
-	if configureForGCP {
-		args = append(args, "--configure-for-gcp")
-	}
-	err = installService(name, exePath, args)
-	if err != nil {
-		return err
-	}
-	log.Printf("Successfully installed service %q.", name)
-	return nil
-}
-
-func installService(name, exePath string, args []string) error {
-	config := mgr.Config{
-		DisplayName: name,
-		Description: "A taskcluster worker that runs on all mainstream platforms",
-		// run as LocalSystem because we call WTSQueryUserToken
-		ServiceStartName: "LocalSystem",
-		ServiceType:      windows.SERVICE_WIN32_OWN_PROCESS | windows.SERVICE_INTERACTIVE_PROCESS,
-		StartType:        mgr.StartAutomatic,
-	}
-	dir := filepath.Dir(exePath)
-	logPath := filepath.Join(dir, "generic-worker-service.log")
-	m, err := mgr.Connect()
-	if err != nil {
-		return err
-	}
-	defer m.Disconnect()
-	s, err := m.CreateService(
-		name,
-		exePath,
-		config,
-		args...,
-	)
-	if err != nil {
-		return err
-	}
-	defer s.Close()
-	log.Printf("Created service %q with exePath %q, logPath %q and args %v",
-		name, exePath, logPath, args)
-
-	// log all events to logfile
-	err = eventlog.Install(
-		name,
-		logPath,
-		false,
-		eventlog.Error|eventlog.Warning|eventlog.Info,
-	)
-	if err != nil {
-		s.Delete()
-		return fmt.Errorf("SetupEventLogSource() failed: %s", err)
-	}
-
-	// start service manually in order to fail fast
-	err = s.Start(args...)
-	if err != nil {
-		return fmt.Errorf("Error starting service %q: %v", name, err)
-	}
-	return nil
-}
-
 func remove(arguments map[string]interface{}) error {
 	switch {
 	case arguments["service"]:
@@ -412,29 +248,6 @@ func remove(arguments map[string]interface{}) error {
 		return removeService(name)
 	}
 	return fmt.Errorf("Unknown remove target - only 'service' is allowed")
-}
-
-func removeService(name string) error {
-	m, err := mgr.Connect()
-	if err != nil {
-		return err
-	}
-	defer m.Disconnect()
-	s, err := m.OpenService(name)
-	if err != nil {
-		return fmt.Errorf("service %s is not installed", name)
-	}
-	defer s.Close()
-	err = s.Delete()
-	if err != nil {
-		return err
-	}
-	err = eventlog.Remove(name)
-	if err != nil {
-		return fmt.Errorf("RemoveEventLogSource() failed: %s", err)
-	}
-	log.Printf("Successfully removed service %q.", name)
-	return nil
 }
 
 func ExePath() (string, error) {
